created the controller for creating new products, some minor changes need to be made.
While the product is created correctly, its image is uploaded to Cloudinary, and the brand ID is properly associated. To perform these tests in Postman, the logic was modified slightly to send booleans and other data that couldn't be sent. This will be changed when testing with a form in React.